move filter functions to OPERATORS
