--- packages/dataviews/src/field-types/utils/filter-greater-than.ts
@@ -1,18 +1,0 @@
-/**
- * Internal dependencies
- */
-import type { NormalizedField } from '../../types';
-
-export default function filterGreaterThan< Item >(
-	item: Item,
-	field: NormalizedField< Item >,
-	filterValue: any
-): boolean {
-	if ( filterValue === undefined ) {
-		return true;
-	}
-
-	const fieldValue = field.getValue( { item } );
-
-	return fieldValue > filterValue;
-}
