--- packages/dataviews/src/field-types/utils/filter-is-all.ts
@@ -1,18 +1,0 @@
-/**
- * Internal dependencies
- */
-import type { NormalizedField } from '../../types';
-
-export default function filterIsAll< Item >(
-	item: Item,
-	field: NormalizedField< Item >,
-	filterValue: any
-): boolean {
-	if ( ! filterValue?.length ) {
-		return true;
-	}
-
-	return filterValue.every( ( value: any ) => {
-		return field.getValue( { item } )?.includes( value );
-	} );
-}
