--- packages/dataviews/src/field-types/utils/filter-not-contains.ts
@@ -1,24 +1,0 @@
-/**
- * Internal dependencies
- */
-import type { NormalizedField } from '../../types';
-
-export default function filterNotContains< Item >(
-	item: Item,
-	field: NormalizedField< Item >,
-	filterValue: any
-): boolean {
-	if ( filterValue === undefined ) {
-		return true;
-	}
-
-	const fieldValue = field.getValue( { item } );
-
-	return (
-		typeof fieldValue === 'string' &&
-		filterValue &&
-		! fieldValue
-			.toLowerCase()
-			.includes( String( filterValue ).toLowerCase() )
-	);
-}
