--- packages/dataviews/src/field-types/utils/filter-starts-with.ts
@@ -1,24 +1,0 @@
-/**
- * Internal dependencies
- */
-import type { NormalizedField } from '../../types';
-
-export default function filterStartsWith< Item >(
-	item: Item,
-	field: NormalizedField< Item >,
-	filterValue: any
-): boolean {
-	if ( filterValue === undefined ) {
-		return true;
-	}
-
-	const fieldValue = field.getValue( { item } );
-
-	return (
-		typeof fieldValue === 'string' &&
-		filterValue &&
-		fieldValue
-			.toLowerCase()
-			.startsWith( String( filterValue ).toLowerCase() )
-	);
-}
