+++ packages/dataviews/src/utils/operators.tsx
@@ -3,12 +3,18 @@
  */
 import { __, sprintf } from '@wordpress/i18n';
 import { createInterpolateElement } from '@wordpress/element';
+import { getDate } from '@wordpress/date';
 import type { ReactElement } from 'react';
 
 /**
  * Internal dependencies
  */
-import type { NormalizedFilter, Operator, Option } from '../types';
+import type {
+	FilterOperator,
+	NormalizedFilter,
+	Operator,
+	Option,
+} from '../types';
 import {
 	OPERATOR_AFTER,
 	OPERATOR_AFTER_INC,
@@ -34,9 +40,45 @@ import {
 	OPERATOR_STARTS_WITH,
 } from '../constants';
 
+import getRelativeDate from '../field-types/utils/get-relative-date';
+
 const filterTextWrappers = {
 	Name: <span className="dataviews-filters__summary-filter-text-name" />,
 	Value: <span className="dataviews-filters__summary-filter-text-value" />,
+};
+
+// Shared operator definition for IS_NONE and IS_NOT_ALL (deprecated).
+const isNoneOperatorDefinition = {
+	/* translators: DataViews operator name */
+	label: __( 'Is none of' ),
+	filterText: ( filter: NormalizedFilter, activeElements: Option[] ) =>
+		createInterpolateElement(
+			sprintf(
+				/* translators: 1: Filter name (e.g. "Author"). 2: Filter value (e.g. "Admin"): "Author is none of: Admin, Editor". */
+				__( '<Name>%1$s is none of: </Name><Value>%2$s</Value>' ),
+				filter.name,
+				activeElements.map( ( element ) => element.label ).join( ', ' )
+			),
+			filterTextWrappers
+		),
+	filter: ( ( item, field, filterValue ) => {
+		if ( ! filterValue?.length ) {
+			return true;
+		}
+
+		const fieldValue = field.getValue( { item } );
+
+		if ( Array.isArray( fieldValue ) ) {
+			return ! filterValue.some( ( fv: any ) =>
+				fieldValue.includes( fv )
+			);
+		} else if ( typeof fieldValue === 'string' ) {
+			return ! filterValue.includes( fieldValue );
+		}
+
+		return false;
+	} ) as FilterOperator< any >,
+	selection: 'multi' as const,
 };
 
 const OPERATORS: {
@@ -46,6 +88,7 @@ const OPERATORS: {
 		filter: NormalizedFilter,
 		activeElements: Option[]
 	) => ReactElement;
+	filter?: FilterOperator< any >;
 	selection: 'single' | 'multi' | 'custom';
 }[] = [
 	{
@@ -64,25 +107,27 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( ! filterValue?.length ) {
+				return true;
+			}
+			const fieldValue = field.getValue( { item } );
+
+			if ( Array.isArray( fieldValue ) ) {
+				return filterValue.some( ( fv: any ) =>
+					fieldValue.includes( fv )
+				);
+			} else if ( typeof fieldValue === 'string' ) {
+				return filterValue.includes( fieldValue );
+			}
+
+			return false;
+		},
 		selection: 'multi',
 	},
 	{
 		name: OPERATOR_IS_NONE,
-		/* translators: DataViews operator name */
-		label: __( 'Is none of' ),
-		filterText: ( filter: NormalizedFilter, activeElements: Option[] ) =>
-			createInterpolateElement(
-				sprintf(
-					/* translators: 1: Filter name (e.g. "Author"). 2: Filter value (e.g. "Admin"): "Author is none of: Admin, Editor". */
-					__( '<Name>%1$s is none of: </Name><Value>%2$s</Value>' ),
-					filter.name,
-					activeElements
-						.map( ( element ) => element.label )
-						.join( ', ' )
-				),
-				filterTextWrappers
-			),
-		selection: 'multi',
+		...isNoneOperatorDefinition,
 	},
 	{
 		name: OPERATOR_IS_ALL,
@@ -100,25 +145,20 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( ! filterValue?.length ) {
+				return true;
+			}
+
+			return filterValue.every( ( value: any ) => {
+				return field.getValue( { item } )?.includes( value );
+			} );
+		},
 		selection: 'multi',
 	},
 	{
 		name: OPERATOR_IS_NOT_ALL,
-		/* translators: DataViews operator name */
-		label: __( 'Is none of' ),
-		filterText: ( filter: NormalizedFilter, activeElements: Option[] ) =>
-			createInterpolateElement(
-				sprintf(
-					/* translators: 1: Filter name (e.g. "Author"). 2: Filter value (e.g. "Admin"): "Author is none of: Admin, Editor". */
-					__( '<Name>%1$s is none of: </Name><Value>%2$s</Value>' ),
-					filter.name,
-					activeElements
-						.map( ( element ) => element.label )
-						.join( ', ' )
-				),
-				filterTextWrappers
-			),
-		selection: 'multi',
+		...isNoneOperatorDefinition,
 	},
 	{
 		name: OPERATOR_BETWEEN,
@@ -137,6 +177,31 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if (
+				! Array.isArray( filterValue ) ||
+				filterValue.length !== 2 ||
+				filterValue[ 0 ] === undefined ||
+				filterValue[ 1 ] === undefined
+			) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			if (
+				typeof fieldValue === 'number' ||
+				fieldValue instanceof Date ||
+				typeof fieldValue === 'string'
+			) {
+				return (
+					fieldValue >= filterValue[ 0 ] &&
+					fieldValue <= filterValue[ 1 ]
+				);
+			}
+
+			return false;
+		},
 		selection: 'custom',
 	},
 	{
@@ -155,6 +220,22 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if (
+				filterValue?.value === undefined ||
+				filterValue?.unit === undefined
+			) {
+				return true;
+			}
+
+			const targetDate = getRelativeDate(
+				filterValue.value,
+				filterValue.unit
+			);
+			const fieldValue = getDate( field.getValue( { item } ) );
+
+			return fieldValue >= targetDate && fieldValue <= new Date();
+		},
 		selection: 'custom',
 	},
 	{
@@ -171,6 +252,22 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if (
+				filterValue?.value === undefined ||
+				filterValue?.unit === undefined
+			) {
+				return true;
+			}
+
+			const targetDate = getRelativeDate(
+				filterValue.value,
+				filterValue.unit
+			);
+			const fieldValue = getDate( field.getValue( { item } ) );
+
+			return fieldValue < targetDate;
+		},
 		selection: 'custom',
 	},
 	{
@@ -187,6 +284,12 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			return (
+				filterValue === field.getValue( { item } ) ||
+				filterValue === undefined
+			);
+		},
 		selection: 'single',
 	},
 	{
@@ -203,6 +306,9 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			return filterValue !== field.getValue( { item } );
+		},
 		selection: 'single',
 	},
 	{
@@ -219,6 +325,15 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return fieldValue < filterValue;
+		},
 		selection: 'single',
 	},
 	{
@@ -237,6 +352,15 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return fieldValue > filterValue;
+		},
 		selection: 'single',
 	},
 	{
@@ -255,6 +379,15 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return fieldValue <= filterValue;
+		},
 		selection: 'single',
 	},
 	{
@@ -273,6 +406,15 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return fieldValue >= filterValue;
+		},
 		selection: 'single',
 	},
 	{
@@ -289,6 +431,16 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const filterDate = getDate( filterValue );
+			const fieldDate = getDate( field.getValue( { item } ) );
+
+			return fieldDate < filterDate;
+		},
 		selection: 'single',
 	},
 	{
@@ -305,6 +457,16 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const filterDate = getDate( filterValue );
+			const fieldDate = getDate( field.getValue( { item } ) );
+
+			return fieldDate > filterDate;
+		},
 		selection: 'single',
 	},
 	{
@@ -323,6 +485,16 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const filterDate = getDate( filterValue );
+			const fieldDate = getDate( field.getValue( { item } ) );
+
+			return fieldDate <= filterDate;
+		},
 		selection: 'single',
 	},
 	{
@@ -341,6 +513,16 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const filterDate = getDate( filterValue );
+			const fieldDate = getDate( field.getValue( { item } ) );
+
+			return fieldDate >= filterDate;
+		},
 		selection: 'single',
 	},
 	{
@@ -357,6 +539,21 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return (
+				typeof fieldValue === 'string' &&
+				filterValue &&
+				fieldValue
+					.toLowerCase()
+					.includes( String( filterValue ).toLowerCase() )
+			);
+		},
 		selection: 'single',
 	},
 	{
@@ -375,6 +572,21 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return (
+				typeof fieldValue === 'string' &&
+				filterValue &&
+				! fieldValue
+					.toLowerCase()
+					.includes( String( filterValue ).toLowerCase() )
+			);
+		},
 		selection: 'single',
 	},
 	{
@@ -391,6 +603,21 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const fieldValue = field.getValue( { item } );
+
+			return (
+				typeof fieldValue === 'string' &&
+				filterValue &&
+				fieldValue
+					.toLowerCase()
+					.startsWith( String( filterValue ).toLowerCase() )
+			);
+		},
 		selection: 'single',
 	},
 	{
@@ -407,6 +634,16 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const filterDate = getDate( filterValue );
+			const fieldDate = getDate( field.getValue( { item } ) );
+
+			return filterDate.getTime() === fieldDate.getTime();
+		},
 		selection: 'single',
 	},
 	{
@@ -423,6 +660,16 @@ const OPERATORS: {
 				),
 				filterTextWrappers
 			),
+		filter( item, field, filterValue ) {
+			if ( filterValue === undefined ) {
+				return true;
+			}
+
+			const filterDate = getDate( filterValue );
+			const fieldDate = getDate( field.getValue( { item } ) );
+
+			return filterDate.getTime() !== fieldDate.getTime();
+		},
 		selection: 'single',
 	},
 ];
